Fixed int with the theta desired

+++ src/cougars_control/src/pid.cpp
@@ -69,7 +69,7 @@ public:
    * @param actual The actual value.
    * @return The output value.
    */
-  int compute(float desired, float actual) {
+  float compute(float desired, float actual) {
 
     // PROPORTIONAL CALCULATIONS
     error = desired - actual;
@@ -95,7 +95,7 @@ public:
     } else if (output < min_output) {
       output = min_output;
     }
-    return (int)output;
+    return (float)output;
   }
 
 private:
